Restructuring & refactoring
